requirements.txt
google-generativeai
ollama
openai
PyMuPDF
streamlit==1.29
toml